requirements.txt
python-dotenv==1.0.1
pydantic==2.10.5
pandas==2.2.3
faker==33.3.1
anthropic==0.42.0
pyyaml==6.0.2
jupyter==1.1.1
ipykernel==6.29.5
black==24.10.0
ruff==0.8.6